+++ changelog.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 
+* 2024-10-15 Breaking change: Events may only be constructed with Builders, Events may have additional trigger conditions, OptionsGraphs indicate when they're running.
 * 2024-10-12 Keywords use builders, addKeyword accepts a builder, VerbBuilder has doNotList() method.
 
 # 7.0.0